Updated to match Zig master, minor playing around with old benchmark (error are good now, thank you Zig core team!)

+++ examples/vcount.zig
@@ -37,7 +37,7 @@ const MyCounter = struct {
         self.count += 1;
     }
 
-    pub fn read(self: *const @This()) usize {
+    pub fn read(self: *const @This()) u32 {
         return self.count;
     }
 };
